refactor: domain 패키지 안의 클래스들을 더 세분화된 패키지로 변경
