+++ src/main/java/blackjack/domain/card/Rank.java
@@ -1,4 +1,4 @@
-package blackjack.domain;
+package blackjack.domain.card;
 
 public enum Rank {
     ACE(1, "A"),
