+++ src/main/java/blackjack/domain/machine/Hit.java
@@ -1,4 +1,4 @@
-package blackjack.domain;
+package blackjack.domain.machine;
 
 import java.util.Arrays;
