+++ src/main/java/blackjack/domain/machine/MatchResult.java
@@ -1,4 +1,4 @@
-package blackjack.domain;
+package blackjack.domain.machine;
 
 import java.util.LinkedHashMap;
 import java.util.Map;
