+++ src/main/java/blackjack/domain/participant/AbstractPlayer.java
@@ -1,4 +1,7 @@
-package blackjack.domain;
+package blackjack.domain.participant;
+
+import blackjack.domain.card.Card;
+import blackjack.domain.card.Deck;
 
 public abstract class AbstractPlayer implements Player {
