+++ src/main/java/blackjack/domain/participant/Player.java
@@ -1,4 +1,7 @@
-package blackjack.domain;
+package blackjack.domain.participant;
+
+import blackjack.domain.card.Card;
+import blackjack.domain.card.Deck;
 
 public interface Player {
