+++ src/test/java/blackjack/domain/AlwaysAscNumberMachine.java
@@ -1,5 +1,6 @@
 package blackjack.domain;
 
+import blackjack.domain.card.CardPickMachine;
 import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
