+++ src/test/java/blackjack/domain/GuestTest.java
@@ -1,5 +1,9 @@
 package blackjack.domain;
 
+import blackjack.domain.card.Card;
+import blackjack.domain.card.Rank;
+import blackjack.domain.card.Suit;
+import blackjack.domain.participant.Guest;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
