+++ src/test/java/blackjack/domain/RankTest.java
@@ -1,5 +1,6 @@
 package blackjack.domain;
 
+import blackjack.domain.card.Rank;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
