+++ src/test/java/blackjack/domain/SuitTest.java
@@ -1,5 +1,6 @@
 package blackjack.domain;
 
+import blackjack.domain.card.Suit;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
